Split live activity code into smaller files
